splitting resolution from adjust-imports plugin
